requirements.txt
asyncio>=3.4.0
numpy>=1.26.0
Pillow>=10.4.0
tqdm>=4.66.0